push

+++ src/main/java/org/Ecommerce/SearchFurniture.java
@@ -48,7 +48,7 @@ public class SearchFurniture implements Search {
         Collections.sort(furnitureList, new Comparator<Product>() {
             @Override
             public int compare(Product o1, Product o2) {
-                return o1.getRating() - o2.getRating();
+                return (int) (o1.getRating() - o2.getRating());
             }
         });
         return furnitureList;
